rename files to avoid linting issues
