+++ packages/cfp-template/template/playwright.config.ts
@@ -12,6 +12,7 @@ const config = {
         port: 3000,
         reuseExistingServer: !process.env.CI,
     },
+    testMatch: "**/*_pw_spec.ts"
 };
 
 module.exports = config;
