+++ packages/cfp-template/template/src/components/welcome/welcome.ts
@@ -1,6 +1,6 @@
 import { customElement, FASTElement } from "@microsoft/fast-element";
-import { welcomeStyles as styles } from "./welcome.styles.js";
-import { welcomeTemplate as template } from "./welcome.template.js";
+import { welcomeStyles as styles } from "./welcome_styles.js";
+import { welcomeTemplate as template } from "./welcome_template.js";
 
 /**
  * A Custom HTML Element.
