+++ packages/cfp-template/template/src/components/welcome/welcome_template.ts
