+++ packages/cfp-template/template/src/index_pw_spec.ts
