+++ packages/fast-cli/.mocharc.json
@@ -2,5 +2,5 @@
     "colors": true,
     "recursive": true,
     "timeout": 100000,
-    "spec": "dist/**/*.spec.js"
+    "spec": "dist/**/*_spec.js"
 }
